requirements.txt
# Application
Django==2.2.10
Pillow
psycopg2==2.7.7
psycopg2-binary==2.7.7
static3
gunicorn
dj-database-url
dj-static
python-decouple
django-widget-tweaks
django-extensions
django-storages
boto3
requests
pytz

# Django Admin
https://github.com/darklow/django-suit/tarball/v2

# Auditlog
git+https://github.com/cosmunsoftwares/django-auditlog.git

# Django Tests
django-test-without-migrations
ipython

# XMLJSON
xmljson

# Django Rest Framework
djangorestframework
itsdangerous
django-cors-headers